+++ ase/db/sqlite.py
@@ -23,7 +23,7 @@ import numpy as np
 
 from ase.data import atomic_numbers
 from ase.db.row import AtomsRow
-from ase.db.core import Database, ops, now, lock, invop
+from ase.db.core import Database, ops, now, lock, invop, parse_selection
 from ase.io.jsonio import encode, decode
 from ase.parallel import parallel_function
 from ase.utils import basestring
@@ -124,6 +124,8 @@ class SQLite3Database(Database, object):
     default = 'NULL'  # used for autoincrement id
     connection = None
     version = None
+    columnnames = [line.split()[0].lstrip()
+                   for line in init_statements[0].splitlines()[1:]]
 
     def _connect(self):
         return sqlite3.connect(self.filename, timeout=600)
@@ -270,8 +272,7 @@ class SQLite3Database(Database, object):
             cur.execute('INSERT INTO systems VALUES ({})'.format(q),
                         values)
         else:
-            q = ', '.join(line.split()[0].lstrip() + '=?'
-                          for line in init_statements[0].splitlines()[2:])
+            q = ', '.join(name + '=?' for name in self.columnnames[1:])
             cur.execute('UPDATE systems SET {} WHERE id=?'.format(q),
                         values + (id,))
 
@@ -288,7 +289,7 @@ class SQLite3Database(Database, object):
         text_key_values = []
         number_key_values = []
         for key, value in key_value_pairs.items():
-            if isinstance(value, (float, int)):
+            if isinstance(value, (float, int, np.bool_)):
                 number_key_values.append([key, float(value), id])
             else:
                 assert isinstance(value, basestring)
@@ -371,7 +372,7 @@ class SQLite3Database(Database, object):
             dct['charges'] = deblob(values[24])
         if values[25] != '{}':
             dct['key_value_pairs'] = decode(values[25])
-        if values[26] != 'null':
+        if len(values) >= 27 and values[26] != 'null':
             dct['data'] = values[26]
 
         return AtomsRow(dct)
@@ -488,7 +489,7 @@ class SQLite3Database(Database, object):
         return sql, args
 
     def _select(self, keys, cmps, explain=False, verbosity=0,
-                limit=None, offset=0, sort=None):
+                limit=None, offset=0, sort=None, include_data=True):
         con = self._connect()
         self._initialize(con)
 
@@ -503,7 +504,8 @@ class SQLite3Database(Database, object):
                         'fmax', 'smax', 'volume', 'mass', 'charge', 'natoms']:
                 sort_table = 'systems'
             else:
-                for dct in self._select(keys + [sort], cmps, limit=1):
+                for dct in self._select(keys + [sort], cmps, limit=1,
+                                        include_data=False):
                     if isinstance(dct['key_value_pairs'][sort], basestring):
                         sort_table = 'text_key_values'
                     else:
@@ -515,8 +517,14 @@ class SQLite3Database(Database, object):
             order = None
             sort_table = None
 
+        if include_data:
+            what = 'systems.*'
+        else:
+            what = ', '.join('systems.' + name
+                             for name in self.columnnames[:26])
+
         sql, args = self.create_select_statement(keys, cmps,
-                                                 sort, order, sort_table)
+                                                 sort, order, sort_table, what)
 
         if explain:
             sql = 'EXPLAIN QUERY PLAN ' + sql
@@ -541,7 +549,7 @@ class SQLite3Database(Database, object):
 
     @parallel_function
     def count(self, selection=None, **kwargs):
-        keys, cmps = self.parse_selection(selection, **kwargs)
+        keys, cmps = parse_selection(selection, **kwargs)
         sql, args = self.create_select_statement(keys, cmps, what='COUNT(*)')
         con = self._connect()
         self._initialize(con)