+++ ase/io/vasp.py
@@ -463,6 +463,7 @@ def read_vasp_xml(filename='vasprun.xml', index=-1):
     atoms_init = None
     calculation = []
     ibz_kpts = None
+    kpt_weights = None
     parameters = OrderedDict()
 
     try:
@@ -483,6 +484,9 @@ def read_vasp_xml(filename='vasprun.xml', index=-1):
 
                     for i, kpt in enumerate(kpts):
                         ibz_kpts[i] = [float(val) for val in kpt.text.split()]
+
+                    kpt_weights = elem.findall('varray[@name="weights"]/v')
+                    kpt_weights = [float(val.text) for val in kpt_weights]
 
                 elif elem.tag == 'parameters':
                     for par in elem.iter():
@@ -626,7 +630,7 @@ def read_vasp_xml(filename='vasprun.xml', index=-1):
             kblocks = step.findall(
                 'eigenvalues/array/set/set/set[@comment="kpoint %d"]' % ikpt)
             if kblocks is not None:
-                for i, kpoint in enumerate(kblocks):
+                for spin, kpoint in enumerate(kblocks):
                     eigenvals = kpoint.findall('r')
                     eps_n = np.zeros(len(eigenvals))
                     f_n = np.zeros(len(eigenvals))
@@ -636,7 +640,8 @@ def read_vasp_xml(filename='vasprun.xml', index=-1):
                         f_n[j] = float(val[1])
                     if len(kblocks) == 1:
                         f_n *= 2
-                    kpoints.append(SinglePointKPoint(1, 0, ikpt, eps_n, f_n))
+                    kpoints.append(SinglePointKPoint(kpt_weights[ikpt - 1],
+                                                     spin, ikpt, eps_n, f_n))
         if len(kpoints) == 0:
             kpoints = None